dream_mecha/requirements.txt
# Discord bot dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0

# Web framework for UI
Flask>=2.3.0
Flask-CORS>=4.0.0

# Database
sqlite3  # Built into Python
# For production: psycopg2-binary>=2.9.0

# Utilities
requests>=2.31.0
aiohttp>=3.8.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0 